split game requirements off of game content
